+++ src/containers/Orders/orders.actions.js
@@ -23,6 +23,9 @@ import {
   MODIFY_ORDER_REQUEST,
   MODIFY_ORDER_SUCCESS,
   MODIFY_ORDER_FAILURE,
+  MODIFY_ORDER_WITH_ZONEID_REQUEST,
+  MODIFY_ORDER_WITH_ZONEID_SUCCESS,
+  MODIFY_ORDER_WITH_ZONEID_FAILURE,
   HANDLE_ADD_ORDER_MODAL_SHOW,
   HANDLE_ADD_ORDER_MODAL_CANCEL,
   HANDLE_MODIFY_ORDER_MODAL_SHOW,
@@ -43,6 +46,8 @@ import {
   ON_CHANGE_COMUNA,
   ON_CHANGE_ADDRESS,
   ON_CHANGE_WEIGHT,
+  ON_CHANGE_LENGTH,
+  ON_CHANGE_HEIGHT,
   ON_CHANGE_VALUE,
   ON_CHANGE_DESCRIPTION,
   ON_CHANGE_QUANTITY,
@@ -50,6 +55,7 @@ import {
   ON_CHANGE_ADD_ORDER_SUCCESS_NUMBER,
   ON_CHANGE_ADD_ORDER_FAIL_NUMBER,
   ON_CHANGE_ISLOADING,
+  ON_CHANGE_WIDTH,
 } from './orders.constants';
 
 export const getOrdersAction = payload => ({ type: GET_ORDERS_REQUEST, payload });
@@ -84,6 +90,10 @@ export const modifyOrdersAction = payload => ({ type: MODIFY_ORDER_REQUEST, payl
 export const modifyOrdersSuccess = payload => ({ type: MODIFY_ORDER_SUCCESS, payload });
 export const modifyOrdersFailure = payload => ({ type: MODIFY_ORDER_FAILURE, payload });
 
+export const modifyOrdersWithZoneIdAction = payload => ({ type: MODIFY_ORDER_WITH_ZONEID_REQUEST, payload });
+export const modifyOrdersWithZoneIdSuccess = payload => ({ type: MODIFY_ORDER_WITH_ZONEID_SUCCESS, payload });
+export const modifyOrdersWithZoneIdFailure = payload => ({ type: MODIFY_ORDER_WITH_ZONEID_FAILURE, payload });
+
 export const handleAddOrderModalShowAction = payload => ({ type: HANDLE_ADD_ORDER_MODAL_SHOW, payload });
 export const handleAddOrderModalCancelAction = payload => ({ type: HANDLE_ADD_ORDER_MODAL_CANCEL, payload });
 
@@ -108,6 +118,9 @@ export const onChangeProvinceAction = payload => ({ type: ON_CHANGE_PROVINCE, pa
 export const onChangeComunaAction = payload => ({ type: ON_CHANGE_COMUNA, payload });
 export const onChangeAddressAction = payload => ({ type: ON_CHANGE_ADDRESS, payload });
 export const onChangeWeightAction = payload => ({ type: ON_CHANGE_WEIGHT, payload });
+export const onChangeLengthAction = payload => ({ type: ON_CHANGE_LENGTH, payload });
+export const onChangeHeightAction = payload => ({ type: ON_CHANGE_HEIGHT, payload });
+export const onChangeWidthAction = payload => ({ type: ON_CHANGE_WIDTH, payload });
 export const onChangeValueAction = payload => ({ type: ON_CHANGE_VALUE, payload });
 export const onChangeDescriptionAction = payload => ({ type: ON_CHANGE_DESCRIPTION, payload });
 export const onChangeQuantityAction = payload => ({ type: ON_CHANGE_QUANTITY, payload });