+++ src/containers/Rules/ruleCharts/level_3_ruleChart.js
@@ -29,16 +29,16 @@ export default function Level_3_RuleChart(props) {
   const positionList = [
     { id: 1, top: '90%', left: '60%' },
     { id: 2, top: '54%', left: '80%' },
-    { id: 3, top: '77%', left: '46%' },
+    { id: 3, top: '77%', left: '49%' },
     { id: 4, top: '43%', left: '70%' },
-    { id: 5, top: '63%', left: '33%' },
+    { id: 5, top: '63%', left: '38%' },
     { id: 6, top: '32%', left: '60%' },
-    { id: 7, top: '49%', left: '21%' },
+    { id: 7, top: '49%', left: '27%' },
     { id: 8, top: '21%', left: '50%' },
-    { id: 9, top: '28%', left: '12%' },
-    { id: 10, top: '10%', left: '43%' },
-    { id: 11, top: '8%', left: '5%' },
-    { id: 12, top: '75%', left: '89%' },
+    { id: 9, top: '34%', left: '16%' },
+    { id: 10, top: '10%', left: '40%' },
+    { id: 11, top: '14%', left: '6%' },
+    { id: 12, top: '75%', left: '90%' },
   ];
 
   const initData = async () => {
@@ -55,11 +55,11 @@ export default function Level_3_RuleChart(props) {
   return (
     <div>
       <div>
-        <div style={{ position: 'absolute', width: 1200, height: 700 }}>
+        <div style={{ position: 'absolute', width: 1500, height: 800 }}>
           <img
             src="./Port_distribution.png"
-            width="1200"
-            height="700"
+            width="1500"
+            height="800"
             style={{ paddingLeft: 200, paddingRight: 100 }}
           />
           {positionList.map(position => {